Added Spells
Added Greater Fade to Fade settings,
Added Dark Archangel, Archangel, Holy Ward, and Ray of Hope.

Ray of Hope has 2 different end messages - Heal or End. End is just when the effect ends, and is blank by default. Heal is when the effect heals you, and includes a an amount replacement for how much it healed.